fix(gcp): add exponential backoff to GCP token refresh retries
