+++ lib/vector-common/src/lib.rs
@@ -94,4 +94,6 @@ pub mod net;
 
 pub mod multiline;
 
+pub mod backoff;
+
 pub mod compression;
